dom/tests/mochitest/bugs/test_bug333983.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=333983
-->
<head>
  <title>Test for Bug 333983</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=333983">Mozilla Bug 333983</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 333983 **/
HTMLElement.prototype.foopy = function () { return "foopy"; }

var foopy = "";
try {
  foopy = document.body.foopy()
} catch (ex) {
  foopy = ex + "";
}

is(foopy, "foopy", "HTMLElement.prototype functions work");
</script>
</pre>
</body>
</html>
